refactor(install-skill): unify docs/hooks into --install-skill
Remove separate --install-docs flag and install-docs.mjs. Extend
skills/install.mjs with GEMINI.md/AGENTS.md merge, Gemini hooks, and
OpenCode plugin per platform registry. Extract install-helpers.mjs.
